Login UI and WideButton Component
Some unnecessary imports are deleted. screenName is changed with screenTitle which is a more appropriate name to use.
LoginScreen UI is created. If LoginScreen is needed, just change the line 10 of App.js.
WideButton component is added. It will be tested and optimized to use as "Log In" and "Sign Up" buttons.