AlbumFollowerScreen, some design updates, new components, and state management
AlbumView component was modified to make it seems modern.

FollowerView component was added.

WideButton component was updated and it is now ready to take its place inside SignUp and Login.

AlbumFollowerScreen was created.

LoginScreen and SignUpScreen were updated. Appropriate states added.

ProfileScreen updated to obtain a profile with passed inputs.

Test data is created to test ProfileScreen and FollowerView. It will be deleted.